refactor: Unspaghetti world_ticker, match_engine, and PromoterDashboard
Three largest files decomposed into focused, single-responsibility modules:

world_ticker.py (2345 → 1850 lines):
  - Extract PlayerActionHandler to player_action_handler.py (563 lines)
    15 action handlers + dispatch table + DRY get_active_contract() helper
  - Extract inter-federation dynamics to inter_federation_service.py (278 lines)
    Market share, talent offers, TV deals, federation momentum

match_engine.py (1819 → 1870 lines, but much cleaner):
  - Extract 30+ magic numbers to match_constants.py (213 lines)
    Named constants for momentum thresholds, probabilities, damage
    modifiers, rating bonuses, tag match values, and finish logic
  - All inline numbers replaced with descriptive constant names

PromoterDashboard.tsx (1174 → 561 lines):
  - Extract usePromoterData custom hook (107 lines)
    11 API calls, loadData, advanceDay, signWrestler consolidated
  - Extract WarRoomTab component (213 lines)
  - Extract FactionsTab component (229 lines)
  - Extract RelationshipsTab component (248 lines)
  - Shared components (AlignmentBadge, HeatBar, RoleBadge) exported
    for reuse across tabs

No behavior changes — pure reorganization. All Python syntax valid,
TypeScript compiles clean, Vite build passes.
