+++ game_service/world_ticker.py
@@ -42,6 +42,8 @@ from models.game_models import (
     WrestlerRelationshipDB, TagTeamDB, TalentOfferDB,
     PromoDB,
 )
+from game_service.player_action_handler import PlayerActionHandler, get_active_contract
+from game_service import inter_federation_service
 
 logger = logging.getLogger(__name__)
 
@@ -233,6 +235,8 @@ class WorldTicker:
 
     def _process_player_actions(self):
         """Process pending player actions."""
+        handler = PlayerActionHandler(self.db, self.world, self._log_event)
+
         pending = self.db.query(PlayerActionDB).filter(
             PlayerActionDB.world_id == self.world.id,
             PlayerActionDB.status == "pending",
@@ -241,7 +245,7 @@ class WorldTicker:
         for action in pending:
             try:
                 action.status = "processing"
-                result = self._execute_player_action(action)
+                result = handler.execute(action)
                 action.status = "completed"
                 action.result = result
                 action.processed_at = datetime.utcnow()
@@ -250,505 +254,6 @@ class WorldTicker:
                 action.status = "failed"
                 action.result = {"error": str(e)}
                 logger.warning(f"Player action {action.id} failed: {e}")
-
-    def _execute_player_action(self, action: PlayerActionDB) -> dict:
-        """Execute a single player action. Returns result dict."""
-        action_type = action.action_type
-        data = action.action_data
-
-        if action_type == "sign_wrestler":
-            return self._action_sign_wrestler(data)
-        elif action_type == "book_show":
-            return self._action_book_show(data)
-        elif action_type == "train":
-            return self._action_train(data)
-        elif action_type == "cut_promo":
-            return self._action_cut_promo(data)
-        elif action_type == "form_stable":
-            return self._action_form_stable(data)
-        elif action_type == "join_stable":
-            return self._action_join_stable(data)
-        elif action_type == "leave_stable":
-            return self._action_leave_stable(data)
-        elif action_type == "dissolve_stable":
-            return self._action_dissolve_stable(data)
-        elif action_type == "assign_manager":
-            return self._action_assign_manager(data)
-        elif action_type == "create_manager":
-            return self._action_create_manager(data)
-        elif action_type == "remove_manager":
-            return self._action_remove_manager(data)
-        elif action_type == "create_storyline":
-            return self._action_create_storyline(data)
-        elif action_type == "advance_storyline":
-            return self._action_advance_storyline(data)
-        elif action_type == "request_match":
-            return self._action_request_match(data)
-        elif action_type == "open_challenge":
-            return self._action_open_challenge(data)
-        else:
-            return {"message": f"Action '{action_type}' acknowledged"}
-
-    def _action_sign_wrestler(self, data: dict) -> dict:
-        """Promoter signs a free agent."""
-        wrestler_id = data.get("wrestler_id")
-        federation_id = data.get("federation_id")
-        salary = data.get("salary_weekly", 1000)
-
-        wrestler = self.db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == wrestler_id
-        ).first()
-        if not wrestler:
-            raise ValueError("Wrestler not found")
-
-        # Check not already signed
-        active_contract = self.db.query(ContractDB).filter(
-            ContractDB.wrestler_id == wrestler_id,
-            ContractDB.status == "active",
-        ).first()
-        if active_contract:
-            raise ValueError("Wrestler already under contract")
-
-        contract = ContractDB(
-            world_id=self.world.id,
-            wrestler_id=wrestler_id,
-            federation_id=federation_id,
-            salary_weekly=salary,
-            start_date=self.world.current_game_date,
-            is_exclusive=True,
-        )
-        self.db.add(contract)
-        self._log_event("signing", f"{wrestler.name} signed with federation", [wrestler_id, federation_id])
-        return {"contract_id": contract.id, "wrestler": wrestler.name}
-
-    def _action_book_show(self, data: dict) -> dict:
-        """Promoter books a show."""
-        show = ShowDB(
-            world_id=self.world.id,
-            federation_id=data["federation_id"],
-            name=data.get("name", "Live Event"),
-            show_type=data.get("show_type", "weekly"),
-            venue=data.get("venue", "Local Arena"),
-            capacity=data.get("capacity", 5000),
-            game_date=data.get("game_date", self.world.current_game_date),
-        )
-        self.db.add(show)
-        self.db.flush()
-        return {"show_id": show.id, "name": show.name, "date": show.game_date}
-
-    def _action_train(self, data: dict) -> dict:
-        """Wrestler trains a stat."""
-        wrestler_id = data.get("wrestler_id")
-        stat_name = data.get("stat", "stamina")
-
-        stats = self.db.query(WrestlerStatsDB).filter(
-            WrestlerStatsDB.wrestler_id == wrestler_id
-        ).first()
-        if not stats:
-            raise ValueError("Wrestler stats not found")
-
-        if not hasattr(stats, stat_name):
-            raise ValueError(f"Invalid stat: {stat_name}")
-
-        current = getattr(stats, stat_name)
-        # Training gain decreases as stat gets higher
-        gain = max(1, random.randint(1, 3) - (current // 40))
-
-        # Mentor bonus (Group 4)
-        try:
-            from game_service.wrestler_lifecycle_service import training_with_mentor
-            gain += training_with_mentor(self.db, wrestler_id, stat_name)
-        except (ValueError, AttributeError) as e:
-            logger.debug("Mentor bonus skipped for %s: %s", wrestler_id, e)
-
-        new_val = min(100, current + gain)
-        setattr(stats, stat_name, new_val)
-
-        # Training fatigues the wrestler slightly
-        wrestler = self.db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == wrestler_id
-        ).first()
-        if wrestler:
-            wrestler.condition = max(0, wrestler.condition - random.randint(2, 5))
-
-        return {"stat": stat_name, "old": current, "new": new_val, "gain": new_val - current}
-
-    def _action_cut_promo(self, data: dict) -> dict:
-        """Wrestler cuts a promo — gains popularity, boosts storyline heat."""
-        wrestler_id = data.get("wrestler_id")
-        target_id = data.get("target_wrestler_id")
-        direction = data.get("direction", "")
-
-        wrestler = self.db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == wrestler_id
-        ).first()
-        if not wrestler:
-            raise ValueError("Wrestler not found")
-
-        stats = self.db.query(WrestlerStatsDB).filter(
-            WrestlerStatsDB.wrestler_id == wrestler_id
-        ).first()
-
-        # Promo quality based on charisma + mic skill
-        charisma = stats.charisma if stats else 50
-        mic = stats.mic_skill if stats else 50
-        base_quality = (charisma + mic) / 2
-        roll = random.randint(-15, 15)
-        quality = max(1, min(100, base_quality + roll))
-
-        # Popularity gain: 1-5 based on quality
-        pop_gain = 1
-        if quality >= 80:
-            pop_gain = random.randint(3, 5)
-        elif quality >= 60:
-            pop_gain = random.randint(2, 4)
-        elif quality >= 40:
-            pop_gain = random.randint(1, 3)
-
-        old_pop = wrestler.popularity
-        wrestler.popularity = min(100, wrestler.popularity + pop_gain)
-        wrestler.morale = min(100, (wrestler.morale or 50) + 2)
-
-        result = {
-            "quality": quality,
-            "popularity_gain": pop_gain,
-            "old_popularity": old_pop,
-            "new_popularity": wrestler.popularity,
-            "direction": direction,
-        }
-
-        # If targeting a rival, boost storyline heat
-        if target_id:
-            sl_svc = _get_storyline_service()
-            existing = sl_svc._find_storyline_between(self.db, wrestler_id, target_id)
-            if existing:
-                heat_boost = 5 if quality >= 60 else 3
-                sl_svc.progress_storyline(
-                    self.db, existing, "promo", heat_boost,
-                    description=f"{wrestler.name} cut a fiery promo targeting their rival!",
-                )
-                result["storyline_heat_boost"] = heat_boost
-
-        self._log_event(
-            "promo",
-            f"{wrestler.name} cuts a promo (quality: {quality}, pop +{pop_gain})",
-            [wrestler_id] + ([target_id] if target_id else []),
-            importance=5,
-        )
-        return result
-
-    # --- Match request / open challenge actions ---
-
-    def _action_request_match(self, data: dict) -> dict:
-        """Player requests a match on the next show for their federation.
-
-        Books the player into an opening slot on the next weekly show.
-        Win = +2-4 popularity, Loss = +1 popularity (exposure value).
-        Cooldown: 1 request per game week (7 ticks).
-        """
-        wrestler_id = data.get("wrestler_id")
-        wrestler = self.db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == wrestler_id
-        ).first()
-        if not wrestler:
-            raise ValueError("Wrestler not found")
-
-        # Find wrestler's federation
-        contract = self.db.query(ContractDB).filter(
-            ContractDB.wrestler_id == wrestler_id,
-            ContractDB.status == "active",
-        ).first()
-        if not contract:
-            raise ValueError("Wrestler has no active contract")
-
-        # Store the request in world metadata for the show booking phase to pick up
-        meta = self.world.world_config or {}
-        match_requests = meta.get("pending_match_requests", [])
-
-        # Check cooldown — only 1 request per 7 days
-        game_date = self.world.current_game_date
-        for req in match_requests:
-            if req.get("wrestler_id") == wrestler_id:
-                last_date = req.get("date", "")
-                try:
-                    days_since = (datetime.strptime(game_date, "%Y-%m-%d") -
-                                  datetime.strptime(last_date, "%Y-%m-%d")).days
-                    if days_since < 7:
-                        return {"message": f"Match request on cooldown ({7 - days_since} days remaining)"}
-                except (ValueError, TypeError):
-                    pass
-
-        # Remove old request for this wrestler if any, add new one
-        match_requests = [r for r in match_requests if r.get("wrestler_id") != wrestler_id]
-        match_requests.append({
-            "wrestler_id": wrestler_id,
-            "federation_id": contract.federation_id,
-            "date": game_date,
-            "type": "request_match",
-        })
-        meta["pending_match_requests"] = match_requests
-        self.world.world_config = meta
-
-        self._log_event(
-            "match_request",
-            f"{wrestler.name} has requested a match on the next show!",
-            [wrestler_id], importance=4,
-        )
-        return {"message": f"{wrestler.name} is booked for a match on the next show", "status": "pending"}
-
-    def _action_open_challenge(self, data: dict) -> dict:
-        """Player issues an open challenge — higher risk/reward than request_match.
-
-        Matched against someone ±10 popularity. Win = +4-6 pop, Loss = +0.
-        Can spark a storyline if player popularity > 40.
-        """
-        wrestler_id = data.get("wrestler_id")
-        wrestler = self.db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == wrestler_id
-        ).first()
-        if not wrestler:
-            raise ValueError("Wrestler not found")
-
-        contract = self.db.query(ContractDB).filter(
-            ContractDB.wrestler_id == wrestler_id,
-            ContractDB.status == "active",
-        ).first()
-        if not contract:
-            raise ValueError("Wrestler has no active contract")
-
-        # Find a suitable opponent: ±10 popularity from same federation
-        roster_contracts = self.db.query(ContractDB).filter(
-            ContractDB.federation_id == contract.federation_id,
-            ContractDB.status == "active",
-            ContractDB.wrestler_id != wrestler_id,
-        ).all()
-
-        candidates = []
-        for c in roster_contracts:
-            opp = self.db.query(GameWrestlerDB).filter(
-                GameWrestlerDB.id == c.wrestler_id,
-                GameWrestlerDB.is_injured == False,
-            ).first()
-            if opp and abs(opp.popularity - wrestler.popularity) <= 15:
-                candidates.append(opp)
-
-        if not candidates:
-            # Fallback: anyone on the roster who isn't injured
-            for c in roster_contracts:
-                opp = self.db.query(GameWrestlerDB).filter(
-                    GameWrestlerDB.id == c.wrestler_id,
-                    GameWrestlerDB.is_injured == False,
-                ).first()
-                if opp:
-                    candidates.append(opp)
-
-        if not candidates:
-            return {"message": "No suitable opponents available for an open challenge"}
-
-        opponent = random.choice(candidates)
-
-        # Store in metadata for show booking
-        meta = self.world.world_config or {}
-        match_requests = meta.get("pending_match_requests", [])
-        match_requests = [r for r in match_requests if r.get("wrestler_id") != wrestler_id]
-        match_requests.append({
-            "wrestler_id": wrestler_id,
-            "federation_id": contract.federation_id,
-            "opponent_id": opponent.id,
-            "date": self.world.current_game_date,
-            "type": "open_challenge",
-        })
-        meta["pending_match_requests"] = match_requests
-        self.world.world_config = meta
-
-        self._log_event(
-            "open_challenge",
-            f"{wrestler.name} issues an open challenge! {opponent.name} answers!",
-            [wrestler_id, opponent.id], importance=6,
-        )
-        return {
-            "message": f"{wrestler.name} issues an open challenge — {opponent.name} has accepted!",
-            "opponent": opponent.name,
-            "opponent_id": opponent.id,
-        }
-
-    # --- Faction / stable actions ---
-
-    def _action_form_stable(self, data: dict) -> dict:
-        """Promoter forms a new stable/faction."""
-        stable_svc = _get_stable_service()
-        name = data.get("name", "New Faction")
-        leader_id = data.get("leader_id")
-        member_ids = data.get("founding_member_ids", [])
-        if not leader_id:
-            raise ValueError("leader_id required")
-        if leader_id not in member_ids:
-            member_ids = [leader_id] + member_ids
-
-        # Find federation from leader's contract
-        contract = self.db.query(ContractDB).filter_by(
-            wrestler_id=leader_id, status="active"
-        ).first()
-        if not contract:
-            raise ValueError("Leader has no active contract")
-
-        stable = stable_svc.create_stable(
-            self.db, self.world.id, contract.federation_id,
-            name=name, leader_id=leader_id,
-            founding_member_ids=member_ids,
-            alignment=data.get("alignment", "heel"),
-            short_name=data.get("short_name"),
-            catchphrase=data.get("catchphrase"),
-            group_finisher_name=data.get("group_finisher_name"),
-            manager_id=data.get("manager_id"),
-            game_date=self.world.current_game_date,
-        )
-        self.events.append(f"Stable formed: {stable.name}")
-        return {"stable_id": stable.id, "name": stable.name}
-
-    def _action_join_stable(self, data: dict) -> dict:
-        """Add a wrestler to an existing stable."""
-        stable_svc = _get_stable_service()
-        stable_id = data.get("stable_id")
-        wrestler_id = data.get("wrestler_id")
-        role = data.get("role", "recruit")
-        if not stable_id or not wrestler_id:
-            raise ValueError("stable_id and wrestler_id required")
-        member = stable_svc.add_member(
-            self.db, stable_id, wrestler_id, role,
-            game_date=self.world.current_game_date,
-        )
-        wrestler = self.db.query(GameWrestlerDB).filter_by(id=wrestler_id).first()
-        self.events.append(f"{wrestler.name if wrestler else wrestler_id} joins stable")
-        return {"member_id": member.id, "role": member.role}
-
-    def _action_leave_stable(self, data: dict) -> dict:
-        """Remove a wrestler from a stable."""
-        stable_svc = _get_stable_service()
-        stable_id = data.get("stable_id")
-        wrestler_id = data.get("wrestler_id")
-        if not stable_id or not wrestler_id:
-            raise ValueError("stable_id and wrestler_id required")
-        result = stable_svc.remove_member(
-            self.db, stable_id, wrestler_id,
-            game_date=self.world.current_game_date,
-        )
-        if not result:
-            raise ValueError("Member not found in stable")
-        return {"removed": True}
-
-    def _action_dissolve_stable(self, data: dict) -> dict:
-        """Dissolve a stable entirely."""
-        stable_svc = _get_stable_service()
-        stable_id = data.get("stable_id")
-        if not stable_id:
-            raise ValueError("stable_id required")
-        from models.game_models import StableDB
-        stable = self.db.query(StableDB).filter_by(id=stable_id).first()
-        if not stable:
-            raise ValueError("Stable not found")
-        stable_svc.dissolve_stable(self.db, stable_id, game_date=self.world.current_game_date)
-        self.events.append(f"Stable dissolved: {stable.name}")
-        return {"dissolved": True, "name": stable.name}
-
-    # --- Manager actions ---
-
-    def _action_assign_manager(self, data: dict) -> dict:
-        """Assign a manager to a wrestler."""
-        mgr_svc = _get_manager_service()
-        manager_id = data.get("manager_id")
-        client_id = data.get("client_wrestler_id")
-        if not manager_id or not client_id:
-            raise ValueError("manager_id and client_wrestler_id required")
-        bond = mgr_svc.assign_manager(
-            self.db, self.world.id, manager_id, client_id,
-            role=data.get("role", "manager"),
-            specialization=data.get("specialization", "all_around"),
-            game_date=self.world.current_game_date,
-        )
-        return {"bond_id": bond.id, "effectiveness": bond.effectiveness}
-
-    def _action_create_manager(self, data: dict) -> dict:
-        """Create a new manager character."""
-        mgr_svc = _get_manager_service()
-        name = data.get("name")
-        if not name:
-            raise ValueError("name required")
-        mgr = mgr_svc.create_manager(
-            self.db, self.world.id, name=name,
-            alignment=data.get("alignment", "heel"),
-            archetype=data.get("archetype", "scheming_manager"),
-            federation_id=data.get("federation_id"),
-            catchphrase=data.get("catchphrase"),
-        )
-        self.events.append(f"Manager created: {mgr.name}")
-        return {"manager_id": mgr.id, "name": mgr.name}
-
-    def _action_remove_manager(self, data: dict) -> dict:
-        """End a manager-client bond."""
-        mgr_svc = _get_manager_service()
-        bond_id = data.get("bond_id")
-        if not bond_id:
-            raise ValueError("bond_id required")
-        result = mgr_svc.remove_manager(
-            self.db, bond_id, game_date=self.world.current_game_date,
-        )
-        if not result:
-            raise ValueError("Bond not found")
-        return {"removed": True}
-
-    # --- Storyline actions ---
-
-    def _action_create_storyline(self, data: dict) -> dict:
-        """Promoter creates a storyline between wrestlers."""
-        sl_svc = _get_storyline_service()
-        wrestler_ids = data.get("wrestler_ids", [])
-        if len(wrestler_ids) < 2:
-            raise ValueError("At least 2 wrestler_ids required")
-        federation_id = data.get("federation_id")
-        if not federation_id:
-            # Infer from first wrestler's contract
-            contract = self.db.query(ContractDB).filter_by(
-                wrestler_id=wrestler_ids[0], status="active"
-            ).first()
-            federation_id = contract.federation_id if contract else None
-        storyline = sl_svc.create_storyline(
-            self.db, self.world.id, federation_id,
-            wrestler_ids=wrestler_ids,
-            storyline_type=data.get("storyline_type", "feud"),
-            name=data.get("name"),
-            description=data.get("description"),
-            game_date=self.world.current_game_date,
-        )
-        self.events.append(f"Storyline created: {storyline.name}")
-        return {"storyline_id": storyline.id, "name": storyline.name}
-
-    def _action_advance_storyline(self, data: dict) -> dict:
-        """Promoter manually advances a storyline's status or heat."""
-        storyline_id = data.get("storyline_id")
-        if not storyline_id:
-            raise ValueError("storyline_id required")
-        storyline = self.db.query(StorylineDB).filter_by(id=storyline_id).first()
-        if not storyline:
-            raise ValueError("Storyline not found")
-
-        new_status = data.get("status")
-        heat_boost = data.get("heat_boost", 0)
-
-        if new_status and new_status in ("brewing", "active", "climax", "resolved"):
-            old_status = storyline.status
-            storyline.status = new_status
-            if new_status == "resolved":
-                storyline.end_date = self.world.current_game_date
-        if heat_boost:
-            storyline.heat = max(0, min(100, storyline.heat + heat_boost))
-
-        return {
-            "storyline_id": storyline.id,
-            "name": storyline.name,
-            "status": storyline.status,
-            "heat": storyline.heat,
-        }
 
     # ------------------------------------------------------------------
     # Phase 2: NPC AI decisions
@@ -1415,10 +920,7 @@ class WorldTicker:
                     morale_change -= 3  # Feeling buried
 
             # Salary relative to popularity
-            contract = self.db.query(ContractDB).filter(
-                ContractDB.wrestler_id == w.id,
-                ContractDB.status == "active",
-            ).first()
+            contract = get_active_contract(self.db, w.id)
             if contract and w.popularity > 60 and contract.salary_weekly < 1500:
                 morale_change -= 2  # Underpaid
 
@@ -1662,230 +1164,24 @@ class WorldTicker:
     def _inter_federation_dynamics(self, game_date: str):
         """Market share, talent poaching, federation momentum."""
         # Daily: update federation momentum
-        self._update_federation_momentum(game_date)
+        inter_federation_service.update_federation_momentum(
+            self.db, self.world, self.events, self._log_event, game_date,
+        )
 
         # Weekly (Sundays): market share redistribution and talent offers
         if get_day_of_week(game_date) == 6:
-            self._redistribute_market_share()
-            self._process_talent_offers(game_date)
-            self._generate_talent_offers(game_date)
-            self._adjust_tv_deals()
-
-    def _update_federation_momentum(self, game_date: str):
-        """Daily federation momentum adjustment."""
-        feds = self.db.query(GameFederationDB).filter(
-            GameFederationDB.world_id == self.world.id,
-            GameFederationDB.is_active == True,
-        ).all()
-
-        for fed in feds:
-            momentum = fed.momentum or 50
-
-            # Recent show quality (check last show)
-            last_show = self.db.query(ShowDB).filter(
-                ShowDB.federation_id == fed.id,
-                ShowDB.is_completed == True,
-                ShowDB.game_date == game_date,
-            ).first()
-
-            if last_show and last_show.overall_rating:
-                if last_show.overall_rating >= 4.0:
-                    momentum += 3
-                elif last_show.overall_rating >= 3.0:
-                    momentum += 1
-                elif last_show.overall_rating < 2.0:
-                    momentum -= 2
-
-            # Roster morale check
-            contracts = self.db.query(ContractDB).filter(
-                ContractDB.federation_id == fed.id,
-                ContractDB.status == "active",
-            ).all()
-            if contracts:
-                wrestler_ids = [c.wrestler_id for c in contracts]
-                wrestlers = self.db.query(GameWrestlerDB).filter(
-                    GameWrestlerDB.id.in_(wrestler_ids),
-                ).all()
-                if wrestlers:
-                    avg_morale = sum(w.morale for w in wrestlers) / len(wrestlers)
-                    if avg_morale < 40:
-                        momentum -= 3  # Scandal/low morale
-
-            # Natural regression toward 50
-            if momentum > 50:
-                momentum -= 1
-            elif momentum < 50:
-                momentum += 1
-
-            fed.momentum = max(0, min(100, momentum))
-
-    def _redistribute_market_share(self):
-        """Redistribute market share based on momentum and show quality."""
-        feds = self.db.query(GameFederationDB).filter(
-            GameFederationDB.world_id == self.world.id,
-            GameFederationDB.is_active == True,
-        ).all()
-
-        if not feds:
-            return
-
-        total_momentum = sum(f.momentum or 50 for f in feds)
-        if total_momentum == 0:
-            return
-
-        for fed in feds:
-            # Market share proportional to momentum
-            target_share = ((fed.momentum or 50) / total_momentum) * 100
-            current = fed.market_share or 0
-            # Gradual shift toward target (10% per week)
-            fed.market_share = round(current + (target_share - current) * 0.1, 1)
-
-    def _generate_talent_offers(self, game_date: str):
-        """NPC feds make talent offers to rivals' wrestlers."""
-        npc_feds = self.db.query(GameFederationDB).filter(
-            GameFederationDB.world_id == self.world.id,
-            GameFederationDB.is_npc == True,
-            GameFederationDB.is_active == True,
-            GameFederationDB.budget > 50000,
-        ).all()
-
-        for fed in npc_feds:
-            if random.random() > 0.15:  # 15% chance per week
-                continue
-
-            # Find targets: popular wrestlers from other feds with low morale
-            targets = self.db.query(GameWrestlerDB).join(ContractDB).filter(
-                GameWrestlerDB.world_id == self.world.id,
-                GameWrestlerDB.is_active == True,
-                GameWrestlerDB.is_npc == True,
-                ContractDB.federation_id != fed.id,
-                ContractDB.status == "active",
-            ).filter(
-                (GameWrestlerDB.popularity > 60) | (GameWrestlerDB.morale < 40)
-            ).all()
-
-            if not targets:
-                continue
-
-            target = random.choice(targets)
-
-            # Check no existing pending offer
-            existing = self.db.query(TalentOfferDB).filter(
-                TalentOfferDB.wrestler_id == target.id,
-                TalentOfferDB.status == "pending",
-            ).first()
-            if existing:
-                continue
-
-            salary = max(1500, target.popularity * 30 + random.randint(-500, 500))
-            expires = advance_game_date(game_date, 14)
-
-            self.db.add(TalentOfferDB(
-                world_id=self.world.id,
-                federation_id=fed.id,
-                wrestler_id=target.id,
-                salary_offered=salary,
-                contract_length_weeks=52,
-                offered_date=game_date,
-                expires_date=expires,
-            ))
-            self._log_event(
-                "talent_offer",
-                f"{fed.short_name or fed.name} makes an offer to {target.name}",
-                [fed.id, target.id], importance=5,
+            inter_federation_service.redistribute_market_share(
+                self.db, self.world, self.events, self._log_event,
             )
-
-    def _process_talent_offers(self, game_date: str):
-        """Process pending talent offers — NPC wrestlers decide."""
-        offers = self.db.query(TalentOfferDB).filter(
-            TalentOfferDB.world_id == self.world.id,
-            TalentOfferDB.status == "pending",
-        ).all()
-
-        for offer in offers:
-            # Expired?
-            if offer.expires_date and offer.expires_date <= game_date:
-                offer.status = "expired"
-                continue
-
-            wrestler = self.db.query(GameWrestlerDB).filter(
-                GameWrestlerDB.id == offer.wrestler_id
-            ).first()
-            if not wrestler or not wrestler.is_npc:
-                continue
-
-            # Current contract
-            current = self.db.query(ContractDB).filter(
-                ContractDB.wrestler_id == wrestler.id,
-                ContractDB.status == "active",
-            ).first()
-
-            accept_chance = 0.1  # Base 10%
-            if wrestler.morale < 30:
-                accept_chance += 0.3
-            if current and offer.salary_offered > current.salary_weekly * 1.5:
-                accept_chance += 0.2
-            if wrestler.popularity > 70:
-                accept_chance -= 0.1  # Popular wrestlers are pickier
-
-            if random.random() < accept_chance:
-                # Accept: end old contract, create new one
-                if current:
-                    current.status = "terminated"
-                    current.end_date = game_date
-
-                new_contract = ContractDB(
-                    world_id=self.world.id,
-                    wrestler_id=wrestler.id,
-                    federation_id=offer.federation_id,
-                    salary_weekly=offer.salary_offered,
-                    start_date=game_date,
-                    is_exclusive=True,
-                )
-                self.db.add(new_contract)
-                offer.status = "accepted"
-
-                fed = self.db.query(GameFederationDB).filter(
-                    GameFederationDB.id == offer.federation_id
-                ).first()
-                fed_name = fed.short_name or fed.name if fed else "Unknown"
-                self._log_event(
-                    "talent_signing",
-                    f"BREAKING: {wrestler.name} signs with {fed_name}!",
-                    [wrestler.id, offer.federation_id], importance=8,
-                )
-                self.events.append(f"{wrestler.name} signs with {fed_name}!")
-                _get_news_service().generate_signing_news(
-                    self.db, self.world.id, wrestler.name, fed_name, game_date,
-                )
-
-                # Momentum shifts
-                if fed:
-                    fed.momentum = min(100, (fed.momentum or 50) + 3)
-            else:
-                offer.status = "rejected"
-
-    def _adjust_tv_deals(self):
-        """Quarterly TV deal adjustments based on performance."""
-        # Only adjust on first Sunday of each quarter-ish (every ~13 weeks)
-        if self.world.current_tick % 91 != 0:
-            return
-
-        feds = self.db.query(GameFederationDB).filter(
-            GameFederationDB.world_id == self.world.id,
-            GameFederationDB.is_active == True,
-        ).all()
-
-        for fed in feds:
-            momentum = fed.momentum or 50
-            if momentum > 70:
-                # Hot fed: TV deal increases
-                increase = fed.tv_deal_value * random.uniform(0.10, 0.20)
-                fed.tv_deal_value += increase
-            elif momentum < 30 and fed.tv_deal_value > 5000:
-                # Cold fed: TV deal shrinks
-                decrease = fed.tv_deal_value * random.uniform(0.05, 0.15)
-                fed.tv_deal_value = max(5000, fed.tv_deal_value - decrease)
+            inter_federation_service.process_talent_offers(
+                self.db, self.world, self.events, self._log_event, game_date,
+            )
+            inter_federation_service.generate_talent_offers(
+                self.db, self.world, self.events, self._log_event, game_date,
+            )
+            inter_federation_service.adjust_tv_deals(
+                self.db, self.world, self.events, self._log_event,
+            )
 
     # ------------------------------------------------------------------
     # Phase 12: Weekly news
